+++ frontend/src/features/auth/LoginForm.svelte
@@ -4,7 +4,7 @@
 	import PasswordField from '$components/input/PasswordField.svelte';
 	import { supabase } from '../../supabase';
 	import { goto } from '$app/navigation';
-	import { Button } from '$components/button';
+	import { Button } from '$components/ui';
 
 	let email: string = '';
 	let password: string = '';
